requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers
coverage
pytest
pytest-cov
pytest-django
daphne
uwsgi
mysqlclient==2.2.7
dnspython==2.7.0
requests
Pillow
jieba
pandas
xlrd
channels
channels-redis
mysqlclient
scikit-learn
